+++ apis/v1/about/search/top.ts
@@ -1,0 +1,5 @@
+import { getRequest } from '@/apis';
+import { AboutSearchResult, SearchParam } from '@/types/search';
+
+export const searchAbout = (params: SearchParam) =>
+  getRequest('/v1/about/search/top', params) as Promise<AboutSearchResult>;
