+++ apis/v2/academics/search/top.ts
@@ -1,0 +1,5 @@
+import { getRequest } from '@/apis';
+import { AcademicsSearchResult, SearchParam } from '@/types/search';
+
+export const searchAcademics = (params: SearchParam) =>
+  getRequest('/v1/academics/search/top', params) as Promise<AcademicsSearchResult>;
